+++ src/main/java/com/lama/dsa/model/order/Coursier.java
@@ -2,19 +2,19 @@ package com.lama.dsa.model.order;
 
 public class Coursier{
 	
-	private int id;
+	private long id;
 	private String name;
 	
-	public Coursier(int id, String name) {
+	public Coursier(long id, String name) {
 		this.id = id;
 		this.name = name;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
